Remove DataFetchers
Follow the conventions of the other sample projects.
