+++ samples/webflux-security/src/main/java/io/spring/sample/graphql/SampleWiring.java
@@ -1,0 +1,49 @@
+package io.spring.sample.graphql;
+
+import java.math.BigDecimal;
+import java.util.Map;
+
+import graphql.schema.idl.RuntimeWiring;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.graphql.boot.RuntimeWiringCustomizer;
+import org.springframework.stereotype.Component;
+
+@Component
+public class SampleWiring implements RuntimeWiringCustomizer {
+
+	final EmployeeService employeeService;
+
+	final SalaryService salaryService;
+
+	public SampleWiring(EmployeeService employeeService, SalaryService salaryService) {
+		this.employeeService = employeeService;
+		this.salaryService = salaryService;
+	}
+
+
+	@Override
+	public void customize(RuntimeWiring.Builder builder) {
+		builder.type("Query", wiringBuilder ->
+			wiringBuilder.dataFetcher("employees", env ->
+				employeeService.getAllEmployees()
+			)
+		);
+		builder.type("Employee", wiringBuilder ->
+			wiringBuilder.dataFetcher("salary",  env -> {
+				Employee employee = env.getSource();
+				return salaryService.getSalaryForEmployee(employee);
+			})
+		);
+		builder.type("Mutation", wiringBuilder ->
+			wiringBuilder.dataFetcher("updateSalary", env -> {
+				Map<String, String> input = env.getArgument("input");
+				String employeeId = input.get("employeeId");
+				BigDecimal newSalary = new BigDecimal(input.get("salary"));
+				salaryService.updateSalary(employeeId, newSalary);
+				return null;
+			})
+		);
+	}
+
+}
